Fix debug info for component instance. Fix reloading debug panel after chasing the debug mode.
https://coveord.atlassian.net/browse/JSUI-1200
